requirements.txt
ensure
joblib
matplotlib
numpy
pandas
python-box
PyYAML
scikit-learn
xgboost
mlflow==2.13.1
dagshub
dotenv
imblearn
Flask
Flask-Cors
-e .
